Add files via upload
In this report, we have built a pricing model for owners and real estate agents that helps in overcoming the problems of overpricing and underpricing by observing the factors that determine the price of the house. Through this methodology, we have identified key factors related to Metro Seattle house pricing such as: view from the windows, number of bathrooms and bedrooms, attractiveness of area, distance to the city center, living space, building condition and age of the building.